+++ frontend/src/pages/Home/Home.tsx
@@ -9,9 +9,10 @@ import ChatBox from '../../components/ChatBox';
 import ChatList from '../../components/ChatList';
 import Sidebar from '../../components/sidebar/Sidebar';
 import GridLayout from 'react-grid-layout';
-import { logout } from '../../api/axiosInstance';
+import { logout } from '../../api/UserInfo';
 import { Dropdown } from 'react-bootstrap';
-import { fetchMessages, fetchConversations, getChatboxes, saveChatbox, resetChatbox } from '../../api/axiosInstance';
+import { fetchMessages, fetchConversations } from '../../api/axiosInstance';
+import { getChatboxes, saveChatbox, resetChatbox } from '../../api/ChatBoxUi';
 import '../../css/Home.css';
 import LoginModal from '../../components/login/LoginModal';
 import { saveSidebarState, loadSidebarState } from '../../utils/sidebarUtils';
