+++ frontend/src/pages/Login.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { loginUser } from '../api/axiosInstance';
+import { loginUser } from '../api/UserInfo';
 import '../css/Login.css';
 import branchImage from '../img/PRlogo2.png';
